requirements.txt
streamlit>=1.35
numpy>=1.26
Pillow>=10.0
matplotlib>=3.8

